+++ tests/Currency.Tests/MoneyTests.cs
@@ -5,6 +5,29 @@ namespace Currency.Tests;
 public class MoneyTests
 {
     [Fact]
+    public void IsDefault_ReturnsTrueForDefaultValue()
+    {
+        Assert.True(default(Money).IsDefault);
+    }
+
+    [Fact]
+    public void IsDefault_ReturnsFalseForInitializedValue()
+    {
+        Assert.False(Money.Zero(CurrencyCode.GBP).IsDefault);
+    }
+
+    [Fact]
+    public void Equality_RemainsSafeForDefaultValues()
+    {
+        var left = default(Money);
+        var right = default(Money);
+
+        Assert.True(left == right);
+        Assert.True(left.Equals(right));
+        _ = left.GetHashCode();
+    }
+
+    [Fact]
     public void Of_AcceptsAmountWithinCurrencyPrecision()
     {
         var money = Money.Of(12.34m, CurrencyCode.GBP);
@@ -57,6 +80,14 @@ public class MoneyTests
     }
 
     [Fact]
+    public void Of_RejectsDefaultCurrency()
+    {
+        var exception = Assert.Throws<ArgumentException>(() => Money.Of(12.34m, default(CurrencyCode)));
+
+        Assert.Contains("initialised", exception.Message);
+    }
+
+    [Fact]
     public void Zero_CreatesZeroMoneyForCurrency()
     {
         var money = Money.Zero(CurrencyCode.USD);
@@ -74,6 +105,14 @@ public class MoneyTests
     }
 
     [Fact]
+    public void FromMinorUnits_RejectsDefaultCurrency()
+    {
+        var exception = Assert.Throws<ArgumentException>(() => Money.FromMinorUnits(1234, default(CurrencyCode)));
+
+        Assert.Contains("initialised", exception.Message);
+    }
+
+    [Fact]
     public void FromMinorUnits_UsesZeroMinorUnitIncrement()
     {
         var money = Money.FromMinorUnits(100, CurrencyCode.JPY);
@@ -87,6 +126,15 @@ public class MoneyTests
         var money = Money.Of(12.34m, CurrencyCode.GBP);
 
         Assert.Equal(1234, money.ToMinorUnits());
+    }
+
+    [Fact]
+    public void ToMinorUnits_RejectsDefaultMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => default(Money).ToMinorUnits());
+
+        Assert.Contains("default", exception.Message);
+        Assert.Contains("Money.Zero", exception.Message);
     }
 
     [Fact]
@@ -107,6 +155,22 @@ public class MoneyTests
     }
 
     [Fact]
+    public void Add_RejectsDefaultMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => default(Money) + Money.Of(10m, CurrencyCode.GBP));
+
+        Assert.Contains("default", exception.Message);
+    }
+
+    [Fact]
+    public void Add_RejectsDefaultOtherMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => Money.Of(10m, CurrencyCode.GBP) + default(Money));
+
+        Assert.Contains("default", exception.Message);
+    }
+
+    [Fact]
     public void Subtract_SubtractsSameCurrencyValues()
     {
         var result = Money.Of(10m, CurrencyCode.GBP) - Money.Of(2.34m, CurrencyCode.GBP);
@@ -121,9 +185,25 @@ public class MoneyTests
     }
 
     [Fact]
+    public void Negate_RejectsDefaultMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => -default(Money));
+
+        Assert.Contains("default", exception.Message);
+    }
+
+    [Fact]
     public void Abs_ReturnsAbsoluteValue()
     {
         Assert.Equal(Money.Of(10m, CurrencyCode.GBP), Money.Of(-10m, CurrencyCode.GBP).Abs());
+    }
+
+    [Fact]
+    public void Abs_RejectsDefaultMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => default(Money).Abs());
+
+        Assert.Contains("default", exception.Message);
     }
 
     [Fact]
@@ -136,6 +216,46 @@ public class MoneyTests
     public void CompareTo_RejectsDifferentCurrencyValues()
     {
         Assert.Throws<InvalidOperationException>(() => Money.Of(10m, CurrencyCode.GBP).CompareTo(Money.Of(10m, CurrencyCode.USD)));
+    }
+
+    [Fact]
+    public void CompareTo_RejectsDefaultMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => default(Money).CompareTo(Money.Of(10m, CurrencyCode.GBP)));
+
+        Assert.Contains("default", exception.Message);
+    }
+
+    [Fact]
+    public void Multiply_RejectsDefaultMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => default(Money).Multiply(2m, CurrencyRoundingPolicy.AwayFromZero()));
+
+        Assert.Contains("default", exception.Message);
+    }
+
+    [Fact]
+    public void Divide_RejectsDefaultMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => default(Money).Divide(2m, CurrencyRoundingPolicy.AwayFromZero()));
+
+        Assert.Contains("default", exception.Message);
+    }
+
+    [Fact]
+    public void Round_RejectsDefaultMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => default(Money).Round(CurrencyRoundingPolicy.AwayFromZero()));
+
+        Assert.Contains("default", exception.Message);
+    }
+
+    [Fact]
+    public void Allocate_RejectsDefaultMoney()
+    {
+        var exception = Assert.Throws<InvalidOperationException>(() => default(Money).Allocate(2));
+
+        Assert.Contains("default", exception.Message);
     }
 
     [Fact]